record.py(直近５回分の出力）
ランキング順の出力ロジック完成。
