+++ backend/app/result.py
@@ -1,9 +1,7 @@
 from flask import Blueprint, request, jsonify
 from model import Result, db, app, Person
-
 from sqlalchemy.orm import sessionmaker
-
-from sqlalchemy import create_engine, Column, String, Integer
+from sqlalchemy import create_engine
 from sqlalchemy.ext.declarative import declarative_base
 
 app.config['JSON_AS_ASCII'] = False
@@ -31,94 +29,37 @@ def get_result():
     ]
     return jsonify(data) 
 
-#直近の5日分出力
-@result_module.route("/user/<user_id>",methods=["POST"])
-def get_my_result(user_id):
-        payload = request.json
-        insert_data = Result(
-            # result_id = payload.get("result_id"),
-            user_id = payload.get("user_id"),
-            accuracy_value = payload.get("accuracy_value"),
-            wpm = payload.get("wpm"),
-            playd_at_date = payload.get("playd_at_date"))
-    
-        users = Result.query.filter_by(user_id=insert_data.user_id).all()
-        
-        my_data = [
-            {
-                "user_id":i.user_id,
-                "accuracy_value":i.accuracy_value,
-                "wpm":i.wpm,
-                "playd_at_date":i.playd_at_date
-            }
-            for i in users
-        ]
-    
-        return jsonify(sorted(my_data, key=lambda x: x['playd_at_date'],reverse=True)[0:5])
 
-#ランキングトップ１０
-# @result_module.route("/1")
-# def get_result_1():
-#     results = Result.query.all()
-#     data = [
-#         {
-#         # "result_id" :i.result_id,
-#         "user_id" : i.user_id,
-#         "accuracy_value" : i.accuracy_value,
-#         "wpm" : i.wpm,
-#         "playd_at_date": i.playd_at_date,
-#         }
-#         for i in results
-#     ]
-#     return jsonify(sorted(data, key=lambda x: x['accuracy_value'],reverse=True)[0:10])
-@result_module.route("/1")
-def get_result_1():
-    results_data = Result.query.all()
+#ランキング1~10まで出力
+
+@result_module.route("/1", methods=["GET"])
+def get_join_result():
+    results = db.session.query(Result, Result.accuracy_value, Result.user_id, Person.user_name, Person.user_id).join(Person,Result.user_id == Person.user_id) 
     data = [
         {
-        # "result_id" :i.result_id,
-        "user_id" : i.user_id,
+        "user_name" : i.user_name,
         "accuracy_value" : i.accuracy_value,
-        "wpm" : i.wpm,
-        "playd_at_date": i.playd_at_date,
         }
-        for i in results_data
+        for i in results
     ]
-    user_table_data = session.query(Person,Result,Person.user_id,Result.accuracy_value)
-    
-    
-    for i in user_table_data:
-        return (
-        "user_id:{}".format(i.user_id),
-        "accuracy_value:{}".format(i.accuracy_value)
-        )
-    
-    # return jsonify(sorted(data, key=lambda x: x['accuracy_value'],reverse=True)[0:10])
-
-
-
-
-
-
+    return jsonify(sorted(data, key=lambda x: x['accuracy_value'],reverse=True)[0:10])
 
 
 #ランキング11~20まで出力
-@result_module.route("/2")
-def get_result_2():
-    results = Result.query.all()
-    # person= Person()
+@result_module.route("/2", methods=["GET"])
+def get_join_result2():
+    results = db.session.query(Result, Result.accuracy_value, Result.user_id, Person.user_name, Person.user_id).join(Person,Result.user_id == Person.user_id) 
     data = [
         {
-        # "result_id" :i.result_id,
-        "user_id" : i.user_id,
+        "user_name" : i.user_name,
         "accuracy_value" : i.accuracy_value,
-        "wpm" : i.wpm,
-        "playd_at_date": i.playd_at_date,
         }
         for i in results
     ]
     return jsonify(sorted(data, key=lambda x: x['accuracy_value'],reverse=True)[10:20])
 
+
+#ゲームの結果を登録
 @result_module.route("",methods=["POST"])
 def post_result():
     payload = request.json
